requirements.txt
pyspark==3.5.0
apache-airflow-providers-amazon
apache-airflow[amazon]